Fixes: NB#270020 - Time format is placed after the date in the phone log
RevBy: Berthold Krevert

Details:

Disable mixing of symbols from lc_messages into the date and
time formats from lc_time *always* if the message locale and the
time locale have use scripts with different direction, i.e.
do not attempt to do this mixing if one of the locales has
a right-to-left script and the other a left-to-right
script.  Mixing for locales with different script
directions almost always gives nonsensical results.