Fixes: NB#269478 - Window Icon is not displaying in the tool bar when rotation changed in the given scenario.
RevBy: Daniel d'Andrada
Details:  The setWindowIconId api is currently not used in the new UI specs and
also by the applications. The current usage of this in widgetsgallery application
is the cause of the bug.  Hence I have removed the same from the menu in the
widgetsgallery application.
The lower level fix is ignored at the moment due to low priority.