Fixes: NB#275709 - Emitting dataChanged after starting animation will result in a crash.
RevBy: Armin Berres.
MergeRequest: 2123

Details: Animations were clearing out the visibleItems array, thus if dataChaged was emitted while the animation was running there was no cell to update.
Fix ensures that if the animation is running dataChanged signals shall not be processed.
As an additional safety measure if the cell that was required to be updated is not found in visibleItems array, the loop will skip it and continue normally.