Fixes: NB#276191 - Object menu is appearing underneath sheet.
RevBy: Daniel d'Andrada
Details: Because of many root elements used by sceneManager, setting zValues
for sceneWindows does not always give expected result. Example of that is
displaying objectMenu on top of sheet. ObjectMenu have bigger zValue
than sheet but because they are added to different root elements (in landscape
orientation) objectMenu is displayed below sheet.

This fix changes sceneManager to always add objectMenus and applicationMenus
to main rootElement (the same sheet is using). Because they need not to
cover statusBar, calculateAvailableSceneRect was updated to take care of that.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2125