Fixes: NB#276512 - Deleting the list view private class while animation is running will cause a crash.
RevBy: John Tapsell, Daniel d'Andrada.
MergeRequest: 2131

Details: When animation class is deleted (in the destructor of list view private), it emits a finished() signal, which will cause the list to recalculate it geometry, while being destroyed.
The fix ensures that isDeleted flag should be set as TRUE, when either destroy() method of the list view privale is called, or its destructor.
And list view private should not ask the view to reprocess geometry due to layout change (when list view private is being deleted), by checking the isDeleted flag.
Additionally the list view private should be disconnected from animations signals in destructor, and animatingItems array should be cleared.