Fixes: NB#277773 - [TASK] Country/ City list: Group header or divider gets messy when setting country and attempting to reselect country again in regional format or time zone country list.
RevBy: John Tapsell

Details:

Improve workaround for index buckets out of the range of the index
buckets used by the current collation locale.

For example, when sorting the Latin strings Á, Aa, Áb in say Greek
locale, a problem occurs because the Greek locale has the buckets

    "[Α Β Γ Δ Ε Ζ Η Θ Ι Κ Λ Μ Ν Ξ Ο Π Ρ Σ Τ Υ Φ Χ Ψ Ω]"

and Latin strings are sorted outside of the range of the
Greek buckets (before the Greek Α).

If some strings ends up outside of the index bucket range
of the locale used for sorting, some workaround is needed,
the current workaround is to use the first character of
the string in this case.

I.e. Á, Aa, Áb get these buckets by using the "first character
if outside of bucket range" workaround:

    -------- Á
    Á
    -------- A
    Aa
    -------- Á
    Áb

which was a bit suboptimal, this commit improves it a lot.