Changes: MTextEdit's text direction for empty text fields is based on input language
RevBy: Pekka Vuorela
MR: 2155
Details: This is libmeegotouch part of the bugfix for NB#252531.

When there is no text in the current new row the cursor position can be wrong.
For Arabic, it should be right aligned. This change uses the actual
input method language to determine what the text direction should initially be.