Fixes: NB#271099 - [TASK] In dialogs, positive action should have blue button.
RevBy: Stanislav Ionascu
MergeRequest: 2183

ModifiedBy: Stanislav Ionascu
Modified: Minor cleanup.

Details:
These changes allow for a custom style name to be set on the MDialog's buttons
by assigning it to the buttons' models. If no style name is set, it will make the
first button highlighted and the remaining will be as before.