Fixes NB#262820 - COREWEB: 'MScalableImage::borders MSliderGroove::paint _q_paintItem QGraphicsScenePrivate::drawItemHelper QGraphicsScenePrivate::draw'
RevBy: Stanislav Ionascu
MergeRequest: 2196

Details:
The MSliderGroove was storing pointers to style's images which would
get deleted on a theme change and this caused the MSliderGroove to
crash because it would use those pointers before they were updated.

These changes make the MSliderGroove use QWeakPointers to check if the
images' pointers are still valid when using them and hence preventing
the crash.