Fixes: NB#278546 - The page handling gets mixed up if a window is minimized and restored while page transitions are on-going
RevBy: Daniel d'Andrada

Details: Page transitions queueing is broken in following scenario:
we have page transition (1) running and request another non animated transition
(2) - it will be queued - and then another animated transition (3) - also will
be queued.
When (1) transition finishes, (2) transition is started, but (3) will stay
in queue until another animated transition is requested.

To fix this I added calling applySceneWindowTransitionsFromPageSwitchQueue()
in push/popPage in case when current transition is not animated. If it is animated,
applySceneWindowTranstionsFromPageSwichQueue will be handled as before on finished
animation signal.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2211