Changes: add some friend class statements, add functions to MLocale
RevBy: Stefan Hundhammer
MR: 2219

Details:
the friend class statements are needed so that we can access
the d pointers of the other classes, so that we can then
give the ML10N::MFoo pointers to the functions where needed
without copying data around.

the new functions in mlocale.h and mlocale.cpp are needed
so that executing of the unit tests does not fail because
of missing functions in the library.