Changes: preparing first Harmattan 1.1.1 release
RevBy: TrustMe
Details: Bump version to 0.23.60~1
Harmattan 1.1.1 and Harmattan 1.1 must share the remainder of the available
version space in 0.23 series, which is [0.23.19, 0.23.99]. So let's give ~40
patch numbers to each.
Harmattan 1.2 already has the 0.24.x range.

If we run out of 0.23.x numbers (very unlikely) we can bump Harmattan 1.2 to 0.26.0
and allocate 0.25.x to Harmattan 1.1.x