Fixes: NB#281175 - Settings; Regional format: Unexpected behavior in changing dates after selecting a different region format
RevBy: Holger Schröder

Details:

The stand-alone day names for Czech were missing, therefore, when
mixing Czech language into the date and time formats of other
languages, the Czech day names were taken from the non-stand-alone
versions which start with lower case.  This was inconsistent.

I use the stand-alone versions of day name and month names when mixing
into the date formats of a different language because the
non-stand-alone versions sometimes have inflections which do not make
any sense at all anymore if they are out of context in a date format
of a different language. Also uppercase is usually better if this
appears in a date format of a different language.