Changes: Limit position indicator height
RevBy: Holger Schröder
MR: 2246

Details: When we scroll too much to -y, indicator
height is counted as -value which requests to
draw pixmap of invalid size. This way we just
paint smallest indicator when we are over limit,
another solution would be to change

indicatorHeight -= indicatorPositionY;

so we would rise the size in -values but I think
it is not valid use case..

Tested with Ut_MPannableViewportScroller.