Changes: Update Ut_MOrientationTracker according to new rotation rules
RevBy: Marcin Miklas
Details:
Previously, windows that are between home screen and the topmost window
would follow the orientation of the topomost window (i.e. current application
window). The new rule states that they should stay put.

Window should be in the same orientation as the user has left it.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2318