Changes: don't cache MOrientationTracker::orientationAngle() value
RevBy: Heikki Holstila
Details:
Don't cache the value, requiring it to be updated whenever one of the
context properties changes.
It's used only to initialize MWindows and MSceneManagers. Thus it's not
called many times in sequence, making the cache approach unnecessary.

This way we also make the update of windows orientations independent
of the computation of the value of MOrientationTracker::orientationAngle().

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2321