Fixes: NB#283251 - [TASK] Windows opened in flat state should follow orientation of the previous top most window
RevBy: Heikki Holstila
Details:
When the device is flat, information provided by Screen.TopEdge is meaningless.
Therefore when new windows are shown while the device is flat they should use the
orientation of the current app. window instead of device orientation (Screen.TopEdge).
E.g.: tapping on a app icon in launcher while the device is lying flat over a table
should make window of that app come up in portrait since homescreen is in portrait and
it's the current window at that moment.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2321