Changes: slider and button - Ungrab gestures when disabled
RevBy: Stanislav Ionascu, Daniel d'Andrada
Details:
Make MSliderView and MButtonSwitchView ungrab pan and swipe gestures
when they get disabled.

This patch is a workaround to avoid the problem explained in
https://projects.maemo.org/bugzilla/show_bug.cgi?id=287082

There seems to be a bug in qt related to detecting gestures and disabled
widgets. Due to this, widgets that grab a gesture and are disabled are
somehow impeding other widgets from getting those gestures.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2331