Fixes: NB#288756 - Panning is not possible if started on a disabled text edit.
RevBy: Daniel d'Andrada
Details: The text edit is grabing the pan gesture by default. When text edit is
disabled the gesture is ignored and is not propagated further.
This fix ensures that text edit grabs the gesture only when it is enabled and
ungrabs it as soon as it's disabled.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2355