Fixes: NB#289461 - MUniqueStringCache must be able to increase the size of the cache
RevBy: Armin Berres
ModifiedBy: Armin Berres
Smaller changed to make this actually work when a cache full situation
has happened.
Additionally only increase once the last chunk is filled by a given
percentage.

Details:
After inserting a new string, an application may increase the size of the cache file
if it's usage is higher than a value, allowing the cache to grow as needed.
The new size of the cache becomes available to the running applications after they
are relaunched.
Without this patch the cache could run full and no applictions can be
started which try to add a new string to the cache.