Changes: Don't call sizeHint explicitly.
RevBy: Michael Hasselmann
Details: sizeHint is only needed to be called by Qt. The values are
cached and using preferredWidth() instead of sizeHint() is therefore
faster. Also, using sizeHint directly may hinder problems caused by
incorrect implementation of it or missing updateGeometry() calls so
in general it is better not to call it ourselves.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2368