Changes: add a function to join a list of strings with localized commas
RevBy: Berthold Krevert

Details:

It seems to be necessary quite often to join some texts like names,
lists of enabled keyboards, days when an alarm clock is enabled,
etc... to a comma separated string. Usually bugs were reported that
the western comma used in the application was not correct for Arabic
or Chinese and sometimes lists which had some strings in Arabic and
others in Latin resulted in broken reordering of some of the strings.
This function should make creating such a comma separated list
easier.

Useful for example for:

NB#288953 - Notification Screen: Contactname is not resolved for Facebook-Chat contact