Changes: Refactoring MSliderIndicator
RevBy: Marcin Miklas
Details:
There's no need to manually call upateGeometry(), resize(). That's the role
of the layout class (in this case a QGraphicsAnchorLayout).

Instead of manually setting the minimum and maximum sizes to be the preferred size
just use the correct size policy: QSizePolicy::Fixed. It has the same meaning.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2375