Fixes: NB#289395 - China Variant: All strings in Mail application are shown in English.
RevBy: Mike Fabian
Details: The ML10N::MLocale class does use abstract config
items to query the system for e.g. the current language.
These items are created by a factory. This factory can either
be the default null factory, or the gconf item factory used
by libmeegotouch. The MLocale in libmeegotouch does set the
gconf factory in the MLocalePrivate constructor, and also
in the MLocale::initMLocale() function.

The problem in this bug is that the fenix-meegotouch package
does use a MCalendar object before it creates a MLocale.
In this case the gconf factory is not set yet when the first
ML10N::MLocale is created. and so that first MLocale picks
up data from the null config items and remembers it for the
runtime of the program.

The fix for the problem is to make sure that MLocale::
initMLocale() is called in every _proxy.cpp i18n class in
libmeegotouch.