Fixes: NB#284847 - MMessageBox::isVisible() and ::isOnDisplay() returns true even if black screensaver is on top of it
RevBy: Daniel d'Andrada, Eduardo Lima Mitev
Details: Newly displayed windows gets onDisplay set to true when show is called
on them. If screen saver is active we expect that onDisplay = false in that
situation, but because of delayed sending of ExitDisplayEvent it happens only
after 1 second.

This patch adds checking if onDisplay status has been previously set by
OnDisplayChangeEvent handlers and if not, it sends ExitDisplayEvent immediately.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2393