Fixes: NB#291403 - Crash in CameraInterface::showCamera called from QML application
RevBy: Daniel d'Andrada

Details: Check if we do have an MApplication instance before
we use more meegotouch functions. As fallback get the
windowId with only qt functions. It would be possible to
also change the "true" branch of the if-statement to use
only qt function calls. But i did not do this to not introduce
the risk of any breakage for the existing working meegotouch
applications.