Fixes: NB#287619 - Presence UI crash appeared when Google is configured with Picasa and Youtube.
RevBy: Daniel d'Andrada
Details:
MSheetStandAloneWindow::hideEvent postpones the disappearance of the sheet to the event
loop. If the sheet disappears before the event loop has had the chance to call it (due
to being dismissed by something else) it results in a crash. The fix adds checks
that the sheet still has a valid scene manager when it is being dismissed from the
event loop.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2415