Fixes: NB#288882 - MMS thread in INBOX not displaying properly when MMS composed with subject full of emoticons
RevBy: John Tapsell, Felipe Erias Morandeira

Details: If the document contains only images but no text, the eliding
might not work and only "..." is shown without any images. The issue occurs
if the height of the images is smaller than the height of the used font.
In this case 'positionForEllipses' gets a negative
y-value which results in a failed QTextDocumentLayout::hitTest().

http://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2429