Fixes: NB#291125, Home screen contents go behind the status bar
RevBy: Daniel d'Andrada
Details: The reason was that relocation (or scroll chain) was not restored
after widget lost focus.

Why not, is because scene does not notify any focus changed events which could
be tracked. Previously, it has been assumed that whenever input method (IM)
hides, it is sufficient to check whether there is a focused widget at
that point in time. This is not the case when IM hiding is initiated from the
IM itself, rather than hiding VKB because of focus lost.

That is, hiding VKB with flick-down gesture triggered relocation before
focus was actually lost. And when focus was lost, nothing notified relocator
about it.

This commit tracks focus lost by installing an event filter to the input widget
and therefore it is now possible to restore scroll chain properly.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2434