Fixes: Slider is set to outdated value on cancelEvent-handler when the initial press didn't hit the slider itself.
RevBy: Daniel d'Andrada
Details: On cancelEvent-handler slider value is set to the value before the press, but that could be outdated
if the press didn't hit the active area (groove or handle) of the slider. Also cancel- and release-feedbacks that
were played although the press were outside of the active area (and thus there were no press-feedback first) are now avoided.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2435