Fixed: NB#293511 - Thai vowel marks are not fully displayed in events view
RevBy: John Tapsell

Details: The tiling code from the richtext-label allowed to draw on
the top and bottom margins but not on the left or right margins. This
is inconsistent to the simple-label behavior where all margins can be
used to e.g. draw too large diacritics or vowel marks.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2441