Fix: NB#264573, MCompleter is covered by VKB
RevBy: Marcin Miklas
Details: Change completer to be positioned around cursor instead of widget,
allowing it also to be above it.

Previously MInputWidgetRelocator positioned completer (implicitly via MTextEdit)
because its area was included in MTextEdit's micro focus rectangle.
This worked as far as there was room inside a pannable viewport below MTextEdit.
To make completer to position itself also above text edit based on its position on
the screen means that it conflicts with relocator and this micro focus hack needs
to be removed.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2469