Changes: Separate cursor rectangle from relocator rectangle.
RevBy: Marcin Miklas

Details: Qt::ImMicroFocus rectangle, the cursor rectangle, was used
previously for relocation purposes. This isn't good enough if we want
to be able to extend MTextEdit's area of interest with MCompleter's area.
New input method query type M::ImRelocatorRectangle is used for the area
of interest, while Qt::ImMicroFocus is left strictly for the cursor.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2469