Fixes: NB#294908 MLabelView - Include the leading() for the last line for height calculations
RevBy: Adrian Yanes
Details:

When I wrote this code I assumed that you would only include the
spacing between the lines, leading(), when it's actually between two
lines.  So that the last line wouldn't include the leading().

This meant, for example, that that minimum sizehint height of an MLabel
is based on the QFontMetrics::height(), rather than  height() +
leading().  However QTextDocument just always includes the leading, even
for a single line of text.

This bug resulted in the minimum height being sometimes a pixel too
small for QTextDocument, resulting in text being unable to be displayed.