Changes: MCompleter shows spinner with async completion model fetches
RevBy: Maksim Kirillov
Details: Completer's behavior is changed so that it can be visible even
with no candidates. This occurs when async fetch is ongoing and completion
prefix exists. In place of the matching candidate, a text label is shown with
the contents retrieved from a new property fetchInProgressLabel.

Required to fix bug NB#288382.
https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2487