Fixes: NB#297788, Empty completer displayed in create message view
RevBy: Adrian Yanes
Details: When contacts model reported it had no ongoing fetch the completer
was unable to hide itself since matched index was still valid.

This will hide completer but it will still flicker in such cases because
contacts model still reports that fetch is ongoing and completer is shown
during this time to indicate fetch to user.