Fixes: NB#295471 - Calendar invitation body is not easily scrolable, scrolls frame by frame
RevBy: Marcin Miklas, Radoslaw Duklas, Armin Berres, Adrian Yanes

Details: The caching mechanism supported only labels with a huge
height. In the scope of NB#295471 the label has a very large width
and the caching fails, which results in the slow fallback where
the text is manually drawn for each frame. The patch extends the
caching mechanism to also support a tile-caching in horizontal
direction.
expand all / collapse all