Fixes: NB#294652 - Settings: In fast scroll, no edge for the characters "ر" and "ز"
RevBy: Adrian Yanes

Details: Characters like "ر" and "ز" have a negative left
bearing which results in a decreased left border. Although this
behavior is usually wanted for a paragraph it looks wrong
in the context of MListIndexToolTip. The patch assures
that for MListIndexToolTip the content-margin is extended
so that the decreased left border gets compensated.